Move vm_errors
