+++ src/vm/errors/mod.rs
@@ -1,0 +1,1 @@
+pub mod vm_errors;
