+++ src/vm/errors/vm_errors.rs
