+++ src/vm/mod.rs
@@ -3,5 +3,5 @@ pub mod decoding;
 pub mod runners;
 pub mod trace;
 pub mod vm_core;
-pub mod vm_errors;
+pub mod errors;
 pub mod vm_memory;
